% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_t2_data.R
\name{delete_t2_data}
\alias{delete_t2_data}
\title{Delete T2 Image Filenames}
\usage{
delete_t2_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "T2"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the T2 images
}
