% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_t2_data.R
\name{download_t2_data}
\alias{download_t2_data}
\title{Download T2 Image Filenames}
\usage{
download_t2_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "T2"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the T2 images
}
