% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_t2_filenames.R
\name{get_t2_filenames}
\alias{get_t2_filenames}
\title{Get T2 Image Filenames}
\usage{
get_t2_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "T2"} so it cannot be specified}
}
\value{
Vector fo filenames
}
\description{
Return the filenames for the T2 images
}
\examples{
get_t2_filenames()
}
